operator: create cell based node watcher metrics type.
The intention is to replace the current one which interacts
with the default promtheus registry directly.
Our intention is to unify operator metrics registration to
using the *metrics.Registry type used in the agent.

These metrics are passed via an interface to the watcher implementations
(i.e. client-go) to register metrics specific to the queue type.

Because of this, we may not want to register a queue type if it is not
needed (i.e. kvstore and/or node_manager).

However, we can find out which metrics types we want at init time
and use the *WithLabels apis to initialize only the needed metrics.

The register functions will now only return a prelabeled metric
for work queue implementation to use.

Signed-off-by: Tom Hadlaw <tom.hadlaw@isovalent.com>