bpf:nodeport:wireguard: encrypt LB traffic in XDP mode
In TC, the to-netdev hook for Wireguard encrypts all overlay traffic,
including all LB-to-remote-backend traffic we generate straight away in
the nodeport codepath.
In XDP, we don't traverse egress program that would apply encryption policy,
and we cannot redirect the packet to the TC to-netdev right away.

This commit adds the capability to encrypt this LB traffic when
WireGuard and NodeEncryption are enabled. The proposed approach acts in the
nodeport codepaths, in particular when the selected backend is remote
and therefore the packet needs to be redirected to the egress interface.
Given we cannot directly apply encryption in the XDP path:

1. push the packet to the TC from-netdev hook, making use of XFER_FLAGS
   to signal which codepath needs to execute for processing the packet.
   While doing so, store in the XDP data_meta the needed values depending
   on the codepath (DSR/Non-DSR, IPv4/IPv6).
2. retrieve apposite data_meta given the XFER_FLAGS in the TC from-netdev
   hook, and store such info into the skb->cb.
3. invoke the remaining portion of the nodeport codepath.
4. given we're in TC, while egressing the packet traverses the to-netdev
   hook, where we apply WireGuard encryption.

Fixes: #38477.

Signed-off-by: Simone Magnani <simone.magnani@isovalent.com>